Former-commit-id: 8443c0b80227cc5b5479bd80379b9772b2fbb45b [formerly 68928d9ed79201c760e1c30eade00625ce93a475]
Former-commit-id: ba3c4744c93c058b0a02b01aa36f3107cf85c39f

+++ backend/RunWithMe/src/main/java/kr/co/rwm/service/ChallengeServiceImpl.java
@@ -29,19 +29,12 @@ public class ChallengeServiceImpl implements ChallengeService {
 
 	@Override
 	public Challenge saveChallenge(ChallengeDto challengeDto) {
-		Challenge challenge = Challenge.builder()
-									.title(challengeDto.getTitle())
-									.content(challengeDto.getContent())
-									.distanceGoal(challengeDto.getDistanceGoal())
-									.distanceCurrent(challengeDto.getDistanceCurrent())
-									.donateGoal(challengeDto.getDonateGoal())
-									.donateCurrent(challengeDto.getDonateCurrent())
-									.startTime(challengeDto.getStartTime())
-									.endTime(challengeDto.getEndTime())
-									.personalDistanceGoal(challengeDto.getPersonalDistanceGoal())
-									.challengeImg(challengeDto.getChallengeImg())
-									.participant(challengeDto.getParticipant())
-									.build();
+		Challenge challenge = Challenge.builder().title(challengeDto.getTitle()).content(challengeDto.getContent())
+				.distanceGoal(challengeDto.getDistanceGoal()).distanceCurrent(challengeDto.getDistanceCurrent())
+				.donateGoal(challengeDto.getDonateGoal()).donateCurrent(challengeDto.getDonateCurrent())
+				.startTime(challengeDto.getStartTime()).endTime(challengeDto.getEndTime())
+				.personalDistanceGoal(challengeDto.getPersonalDistanceGoal())
+				.challengeImg(challengeDto.getChallengeImg()).participant(challengeDto.getParticipant()).build();
 		return challengeRepository.save(challenge);
 	}
 
@@ -75,13 +68,13 @@ public class ChallengeServiceImpl implements ChallengeService {
 	@Override
 	public int deleteChallenge(int challengeId) {
 		Optional<Challenge> challenge = challengeRepository.findByChallengeId(challengeId);
-		if(!challenge.isPresent()) {
+		if (!challenge.isPresent()) {
 			return -1;
 		}
-		
-		if(challenge.get().getParticipant()>0) {
+
+		if (challenge.get().getParticipant() > 0) {
 			return 0;
-		}else {
+		} else {
 			challengeRepository.deleteByChallengeId(challenge.get().getChallengeId());
 			return 1;
 		}
@@ -112,8 +105,9 @@ public class ChallengeServiceImpl implements ChallengeService {
 		// 진행중
 		List<Challenge> ingChallenges = new ArrayList<Challenge>();
 		List<Challenge> challenges = challengeRepository.findAllByEndTimeGreaterThanEqual(endTime);
-		for(Challenge challenge: challenges) {
-			if(challenge.getStartTime().isAfter(endTime)) continue;
+		for (Challenge challenge : challenges) {
+			if (challenge.getStartTime().isAfter(endTime))
+				continue;
 			ingChallenges.add(challenge);
 		}
 		return ingChallenges;
@@ -125,7 +119,7 @@ public class ChallengeServiceImpl implements ChallengeService {
 		List<Challenge> challenges = challengeRepository.findAllByStartTimeGreaterThan(startTime);
 		return challenges;
 	}
-	
+
 	@Override
 	public List<Challenge> findAllChallengeLessThanEndTime(LocalDateTime endTime) {
 		// 끝남
@@ -135,46 +129,67 @@ public class ChallengeServiceImpl implements ChallengeService {
 
 	@Override
 	public ChallengeUser participateChallenge(int challengeId, int donation, int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		
-		Challenge challenge = challengeRepository.findByChallengeId(challengeId).get();
-		challenge.setParticipant(challenge.getParticipant()+1);
-		challenge.setDonateCurrent(challenge.getDonateCurrent()+donation);
-		challengeRepository.save(challenge);
-		
-		ChallengeUser challengeUser = ChallengeUser.builder()
-												.userId(user)
-												.donation(donation)
-												.challengeId(challenge)
-												.accDistance(0)
-												.build();
-		return challengeUserRepository.save(challengeUser);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			Optional<Challenge> challenges = challengeRepository.findByChallengeId(challengeId);
+			if (challenges.isPresent()) {
+				Challenge challenge = challenges.get();
+				challenge.setParticipant(challenge.getParticipant() + 1);
+				challenge.setDonateCurrent(challenge.getDonateCurrent() + donation);
+				challengeRepository.save(challenge);
+
+				ChallengeUser challengeUser = ChallengeUser.builder().userId(user).donation(donation)
+						.challengeId(challenge).accDistance(0).build();
+				return challengeUserRepository.save(challengeUser);
+			} else {
+				return null;
+			}
+		} else {
+			return null;
+		}
+
 	}
 
 	@Transactional
 	@Override
 	public Challenge cancelChallenge(int challengeId, int donation, int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		
-		Challenge challenge = challengeRepository.findByChallengeId(challengeId).get();
-		challenge.setParticipant(challenge.getParticipant()-1);
-		challenge.setDonateCurrent(challenge.getDonateCurrent()-donation);
-		challengeRepository.save(challenge);
-		
-		ChallengeUser challengeUser = challengeUserRepository.findByUserIdAndChallengeId(user, challenge);
-		challengeUserRepository.delete(challengeUser);
-		return challenge;
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			Optional<Challenge> challenges = challengeRepository.findByChallengeId(challengeId);
+			if (challenges.isPresent()) {
+				Challenge challenge = challenges.get();
+				challenge.setParticipant(challenge.getParticipant() - 1);
+				challenge.setDonateCurrent(challenge.getDonateCurrent() - donation);
+				challengeRepository.save(challenge);
+
+				ChallengeUser challengeUser = challengeUserRepository.findByUserIdAndChallengeId(user, challenge);
+				challengeUserRepository.delete(challengeUser);
+				return challenge;
+			} else {
+				return null;
+			}
+		} else {
+			return null;
+		}
+
 	}
 
 	@Override
 	public List<Integer> findByChallengeUserList(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUser = challengeUserRepository.findAllByUserId(user);
-		List<Integer> challengeIds = new ArrayList<Integer>();
-		for (ChallengeUser cu : challengeUser) {
-			challengeIds.add(cu.getChallengeId().getChallengeId());
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUser = challengeUserRepository.findAllByUserId(user);
+			List<Integer> challengeIds = new ArrayList<Integer>();
+			for (ChallengeUser cu : challengeUser) {
+				challengeIds.add(cu.getChallengeId().getChallengeId());
+			}
+			return challengeIds;
+		} else {
+			return null;
 		}
-		return challengeIds;
 	}
 
 	@Override
@@ -183,19 +198,23 @@ public class ChallengeServiceImpl implements ChallengeService {
 		double newDistance;
 		LocalDateTime today = LocalDateTime.now();
 		for (ChallengeUser cu : challengeUsers) {
-			if ((cu.getChallengeId().getEndTime()).isBefore(today) 
-					|| cu.getChallengeId().getStartTime().isAfter(today))
+			if ((cu.getChallengeId().getEndTime()).isBefore(today) || cu.getChallengeId().getStartTime().isAfter(today))
 				continue;
-			
+
 			// challenge update
-			Challenge challenge = challengeRepository.findByChallengeId(cu.getChallengeId().getChallengeId()).get();
-			challenge.setDistanceCurrent(challenge.getDistanceCurrent()+accDistance);
-			challengeRepository.save(challenge);
-			
-			// challenge_user update
-			newDistance = cu.getAccDistance() + accDistance;
-			cu.setAccDistance(newDistance);
-			challengeUserRepository.save(cu); // update challenge user
+			Optional<Challenge> challenges = challengeRepository
+					.findByChallengeId(cu.getChallengeId().getChallengeId());
+			if (challenges.isPresent()) {
+				Challenge challenge = challenges.get();
+				challenge.setDistanceCurrent(challenge.getDistanceCurrent() + accDistance);
+				challengeRepository.save(challenge);
+
+				// challenge_user update
+				newDistance = cu.getAccDistance() + accDistance;
+				cu.setAccDistance(newDistance);
+				challengeUserRepository.save(cu); // update challenge user
+
+			}
 		}
 	}
 
@@ -207,7 +226,7 @@ public class ChallengeServiceImpl implements ChallengeService {
 		yesterday = yesterday.withHour(23).withMinute(59).withSecond(59).withNano(0);
 
 		List<Challenge> challenges = challengeRepository.findAllByEndTime(yesterday);
-		
+
 		double personalGoal;
 		for (Challenge c : challenges) {
 			personalGoal = c.getPersonalDistanceGoal();
@@ -227,131 +246,177 @@ public class ChallengeServiceImpl implements ChallengeService {
 
 	@Override
 	public ChallengeUser findChallengeUserByUserIdAndChallengeId(int userId, Challenge challenge) {
-		User user = userRepository.findByUserId(userId).get();
-		ChallengeUser challengeUser = challengeUserRepository.findByUserIdAndChallengeId(user, challenge);
-		return challengeUser;
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			ChallengeUser challengeUser = challengeUserRepository.findByUserIdAndChallengeId(user, challenge);
+			return challengeUser;
+		} else {
+			return null;
+		}
 	}
 
 	@Override
 	public List<ChallengeUser> findAllChallengeUserByUserIdIng(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
-		List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>(); 
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			// 끝나는 시간이 오늘보다 나중이고, 시작시간이 오늘보다 이전인 경우
-			if ((cu.getChallengeId().getEndTime()).isAfter(today) 
-					&& cu.getChallengeId().getStartTime().isBefore(today)) {
-				challengeUsers.add(cu);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
+			List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>();
+
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				// 끝나는 시간이 오늘보다 나중이고, 시작시간이 오늘보다 이전인 경우
+				if ((cu.getChallengeId().getEndTime()).isAfter(today)
+						&& cu.getChallengeId().getStartTime().isBefore(today)) {
+					challengeUsers.add(cu);
+				}
 			}
+			return challengeUsers;
+		} else {
+			return null;
 		}
-		return challengeUsers;
 	}
+
 	// 참여 안하고 있는 챌린지
 	@Override
 	public List<ChallengeUser> findAllNonChallengeUserByUserIdIng(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserIdNot(user);
-		List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>(); 
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			// 끝나는 시간이 오늘보다 나중이고, 시작시간이 오늘보다 이전인 경우
-			if ((cu.getChallengeId().getEndTime()).isAfter(today) 
-					&& cu.getChallengeId().getStartTime().isBefore(today)) {
-				challengeUsers.add(cu);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserIdNot(user);
+			List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>();
+
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				// 끝나는 시간이 오늘보다 나중이고, 시작시간이 오늘보다 이전인 경우
+				if ((cu.getChallengeId().getEndTime()).isAfter(today)
+						&& cu.getChallengeId().getStartTime().isBefore(today)) {
+					challengeUsers.add(cu);
+				}
 			}
+			return challengeUsers;
+
+		} else {
+			return null;
 		}
-		return challengeUsers;
 	}
 
 	@Override
 	public List<ChallengeUser> findAllChallengeUserByUserIdComingSoon(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
-		List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>(); 
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			// 시작시간이 오늘보다 나중일 때
-			if (cu.getChallengeId().getStartTime().isAfter(today)) {
-				challengeUsers.add(cu);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
+			List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>();
+
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				// 시작시간이 오늘보다 나중일 때
+				if (cu.getChallengeId().getStartTime().isAfter(today)) {
+					challengeUsers.add(cu);
+				}
 			}
+			return challengeUsers;
+		} else {
+			return null;
 		}
-		return challengeUsers;
+
 	}
+
 	// 참여 안하고 있는것
 	@Override
 	public List<ChallengeUser> findAllNonChallengeUserByUserIdComingSoon(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserIdNot(user);
-		List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>(); 
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			// 시작시간이 오늘보다 나중일 때
-			if (cu.getChallengeId().getStartTime().isAfter(today)) {
-				challengeUsers.add(cu);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserIdNot(user);
+			List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>();
+
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				// 시작시간이 오늘보다 나중일 때
+				if (cu.getChallengeId().getStartTime().isAfter(today)) {
+					challengeUsers.add(cu);
+				}
 			}
+			return challengeUsers;
+
+		} else {
+			return null;
 		}
-		return challengeUsers;
+
 	}
 
 	@Override
 	public List<ChallengeUser> findAllChallengeUserByUserIdEnd(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
-		List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>(); 
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			// 끝나는 시간이 오늘보다 이전일때
-			if ((cu.getChallengeId().getEndTime()).isBefore(today)) {
-				challengeUsers.add(cu);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
+			List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>();
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				// 끝나는 시간이 오늘보다 이전일때
+				if ((cu.getChallengeId().getEndTime()).isBefore(today)) {
+					challengeUsers.add(cu);
+				}
 			}
+			return challengeUsers;
+		} else {
+			return null;
 		}
-		return challengeUsers;
+
 	}
+
 	// 참여 안하고 있는 것
 	@Override
 	public List<ChallengeUser> findAllNonChallengeUserByUserIdEnd(int userId) {
-		User user = userRepository.findByUserId(userId).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserIdNot(user);
-		List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>(); 
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			// 끝나는 시간이 오늘보다 이전일때
-			if ((cu.getChallengeId().getEndTime()).isBefore(today)) {
-				challengeUsers.add(cu);
+		Optional<User> users = userRepository.findByUserId(userId);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserIdNot(user);
+			List<ChallengeUser> challengeUsers = new ArrayList<ChallengeUser>();
+
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				// 끝나는 시간이 오늘보다 이전일때
+				if ((cu.getChallengeId().getEndTime()).isBefore(today)) {
+					challengeUsers.add(cu);
+				}
 			}
+			return challengeUsers;
+
+		} else {
+			return null;
 		}
-		return challengeUsers;
+
 	}
 
 	@Override
 	public void deleteAllChallengeUserByUserEmail(String userEmail) {
-		User user = userRepository.findByUserEmail(userEmail).get();
-		List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
-		
-		LocalDateTime today = LocalDateTime.now();
-		today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
-		for(ChallengeUser cu: challengeUserList) {
-			Challenge challenge = challengeRepository.findByChallengeId(cu.getChallengeId().getChallengeId()).get();
-			if (challenge.getEndTime().isBefore(today)) continue;
-			
-			challenge.setParticipant(challenge.getParticipant()-1);
-			challenge.setDistanceCurrent(challenge.getDistanceCurrent()-cu.getAccDistance());
-			challengeRepository.save(challenge);
-		}
-		
-	}
+		Optional<User> users = userRepository.findByUserEmail(userEmail);
+		if (users.isPresent()) {
+			User user = users.get();
+			List<ChallengeUser> challengeUserList = challengeUserRepository.findAllByUserId(user);
 
+			LocalDateTime today = LocalDateTime.now();
+			today = today.withHour(23).withMinute(59).withSecond(59).withNano(0);
+			for (ChallengeUser cu : challengeUserList) {
+				Challenge challenge = challengeRepository.findByChallengeId(cu.getChallengeId().getChallengeId()).get();
+				if (challenge.getEndTime().isBefore(today))
+					continue;
+
+				challenge.setParticipant(challenge.getParticipant() - 1);
+				challenge.setDistanceCurrent(challenge.getDistanceCurrent() - cu.getAccDistance());
+				challengeRepository.save(challenge);
+			}
+		}
+	}
 }
